refactor(execution): remove compatibility aliases and shared event helper (#273)
