+++ tests/execution/test_cancellation.py
@@ -86,7 +86,7 @@ async def test_cancel_interrupts_running_execute_and_keeps_queue_open(caplog):
         "metric=a2a_cancel_abort_success_total" in record.message for record in caplog.records
     )
     assert any("metric=a2a_cancel_duration_ms" in record.message for record in caplog.records)
-    assert executor._sessions.get(("user-1", "context-A")) is None
+    assert executor._session_manager._sessions.get(("user-1", "context-A")) is None
     assert ("task-1", "context-A") not in executor._running_requests
     assert ("task-1", "context-A") not in executor._running_stop_events
     assert ("task-1", "context-A") not in executor._running_identities
@@ -261,7 +261,7 @@ async def test_cancel_waiting_for_session_lock_does_not_abort_other_generation()
     executor = OpencodeAgentExecutor(client, streaming_enabled=False)
     prelocked_session_lock = asyncio.Lock()
     await prelocked_session_lock.acquire()
-    executor._session_locks["session-4"] = prelocked_session_lock
+    executor._session_manager._session_locks["session-4"] = prelocked_session_lock
 
     execute_context = make_request_context_mock(
         task_id="task-4",
@@ -274,7 +274,7 @@ async def test_cancel_waiting_for_session_lock_does_not_abort_other_generation()
 
     try:
         for _ in range(20):
-            if executor._sessions.get(("user-4", "context-D")) == "session-4":
+            if executor._session_manager._sessions.get(("user-4", "context-D")) == "session-4":
                 break
             await asyncio.sleep(0.01)
